requirements.txt
django
django-taggit
django-polymorphic
django-extensions
django-sekizai
django-reversion
django-grappelli
PIL

